Remove experimental test code
