+++ src/components/ha-clickable-list-item.ts
@@ -1,6 +1,4 @@
-import "./ha-clickable-ripple";
 import { ListItem } from "@material/mwc-list/mwc-list-item";
-// import { ListItemBase } from "@material/mwc-list/mwc-list-item-base";
 import { css, CSSResult, customElement } from "lit-element";
 import { html } from "lit-html";
 
@@ -26,12 +24,10 @@ export class HaClickableListItem extends ListItem {
           padding-left: 0px;
           padding-right: 0px;
         }
-
         :host([graphic="avatar"]:not([twoLine])),
         :host([graphic="icon"]:not([twoLine])) {
           height: 48px;
         }
-
         a {
           width: 100%;
           height: 100%;
