+++ src/components/ha-sidebar.ts
@@ -1,10 +1,20 @@
-import "./ha-sidebar-panel-list";
 import "./ha-clickable-list-item";
-import "@material/mwc-list/mwc-list-item";
-import "@material/mwc-list/mwc-list";
+import "./ha-icon";
+import "./ha-menu-button";
+import "./ha-svg-icon";
+import "./user/ha-user-badge";
 import "@material/mwc-button/mwc-button";
 import "@material/mwc-icon-button";
-import { mdiBell, mdiMenu, mdiMenuOpen } from "@mdi/js";
+import { List } from "@material/mwc-list";
+import {
+  mdiBell,
+  mdiCellphoneCog,
+  mdiClose,
+  mdiMenu,
+  mdiMenuOpen,
+  mdiPlus,
+  mdiViewDashboard,
+} from "@mdi/js";
 import {
   css,
   CSSResult,
@@ -15,26 +25,143 @@ import {
   LitElement,
   property,
   PropertyValues,
+  query,
 } from "lit-element";
 import { classMap } from "lit-html/directives/class-map";
+import { guard } from "lit-html/directives/guard";
+import memoizeOne from "memoize-one";
+import { LocalStorage } from "../common/decorators/local-storage";
 import { fireEvent } from "../common/dom/fire_event";
 import { computeDomain } from "../common/entity/compute_domain";
+import { compare } from "../common/string/compare";
 import { computeRTL } from "../common/util/compute_rtl";
 import { ActionHandlerDetail } from "../data/lovelace";
 import {
   PersistentNotification,
   subscribeNotifications,
 } from "../data/persistent_notification";
-import { getExternalConfig } from "../external_app/external_config";
+import {
+  ExternalConfig,
+  getExternalConfig,
+} from "../external_app/external_config";
 import { actionHandler } from "../panels/lovelace/common/directives/action-handler-directive";
 import { haStyleScrollbar } from "../resources/styles";
-import type { HomeAssistant } from "../types";
-import "./ha-icon";
-import "./ha-menu-button";
-import "./ha-svg-icon";
-import "./user/ha-user-badge";
+import type { HomeAssistant, PanelInfo } from "../types";
+import { ListItem } from "@material/mwc-list/mwc-list-item";
+
+const SHOW_AFTER_SPACER = ["config", "developer-tools", "hassio"];
 
 const SUPPORT_SCROLL_IF_NEEDED = "scrollIntoViewIfNeeded" in document.body;
+
+const SORT_VALUE_URL_PATHS = {
+  map: 1,
+  logbook: 2,
+  history: 3,
+  "developer-tools": 9,
+  hassio: 10,
+  config: 11,
+};
+
+const panelSorter = (
+  reverseSort: string[],
+  defaultPanel: string,
+  a: PanelInfo,
+  b: PanelInfo
+) => {
+  const indexA = reverseSort.indexOf(a.url_path);
+  const indexB = reverseSort.indexOf(b.url_path);
+  if (indexA !== indexB) {
+    if (indexA < indexB) {
+      return 1;
+    }
+    return -1;
+  }
+  return defaultPanelSorter(defaultPanel, a, b);
+};
+
+const defaultPanelSorter = (
+  defaultPanel: string,
+  a: PanelInfo,
+  b: PanelInfo
+) => {
+  // Put all the Lovelace at the top.
+  const aLovelace = a.component_name === "lovelace";
+  const bLovelace = b.component_name === "lovelace";
+
+  if (a.url_path === defaultPanel) {
+    return -1;
+  }
+  if (b.url_path === defaultPanel) {
+    return 1;
+  }
+
+  if (aLovelace && bLovelace) {
+    return compare(a.title!, b.title!);
+  }
+  if (aLovelace && !bLovelace) {
+    return -1;
+  }
+  if (bLovelace) {
+    return 1;
+  }
+
+  const aBuiltIn = a.url_path in SORT_VALUE_URL_PATHS;
+  const bBuiltIn = b.url_path in SORT_VALUE_URL_PATHS;
+
+  if (aBuiltIn && bBuiltIn) {
+    return SORT_VALUE_URL_PATHS[a.url_path] - SORT_VALUE_URL_PATHS[b.url_path];
+  }
+  if (aBuiltIn) {
+    return -1;
+  }
+  if (bBuiltIn) {
+    return 1;
+  }
+  // both not built in, sort by title
+  return compare(a.title!, b.title!);
+};
+
+const computePanels = memoizeOne(
+  (
+    panels: HomeAssistant["panels"],
+    defaultPanel: HomeAssistant["defaultPanel"],
+    panelsOrder: string[],
+    hiddenPanels: string[]
+  ): [PanelInfo[], PanelInfo[]] => {
+    if (!panels) {
+      return [[], []];
+    }
+
+    const beforeSpacer: PanelInfo[] = [];
+    const afterSpacer: PanelInfo[] = [];
+
+    Object.values(panels).forEach((panel) => {
+      if (
+        hiddenPanels.includes(panel.url_path) ||
+        (!panel.title && panel.url_path !== defaultPanel)
+      ) {
+        return;
+      }
+      (SHOW_AFTER_SPACER.includes(panel.url_path)
+        ? afterSpacer
+        : beforeSpacer
+      ).push(panel);
+    });
+
+    const reverseSort = [...panelsOrder].reverse();
+
+    beforeSpacer.sort((a, b) => panelSorter(reverseSort, defaultPanel, a, b));
+    afterSpacer.sort((a, b) => panelSorter(reverseSort, defaultPanel, a, b));
+
+    return [beforeSpacer, afterSpacer];
+  }
+);
+
+const isListItem = (element: Element): element is ListItem =>
+  element.hasAttribute("mwc-list-item");
+
+const isNodeElement = (node: Node): node is Element =>
+  node.nodeType === Node.ELEMENT_NODE;
 
 let Sortable;
 
@@ -50,11 +177,21 @@ class HaSidebar extends LitElement {
 
   @property({ type: Boolean }) public editMode = false;
 
+  @internalProperty() private _externalConfig?: ExternalConfig;
+
   @internalProperty() private _notifications?: PersistentNotification[];
 
   // property used only in css
   // @ts-ignore
   @property({ type: Boolean, reflect: true }) public rtl = false;
+
+  @internalProperty() private _renderEmptySortable = false;
+
+  @query("mwc-list.ha-scrollbar", false)
+  private _standardPanelList?: List;
+
+  @query("mwc-list.utility-panels", false)
+  private _utilityPanelList?: List;
 
   private _mouseLeaveTimeout?: number;
 
@@ -62,33 +199,29 @@ class HaSidebar extends LitElement {
 
   private _recentKeydownActiveUntil = 0;
 
+  // @ts-ignore
+  @LocalStorage("sidebarPanelOrder", true, {
+    attribute: false,
+  })
+  private _panelOrder: string[] = [];
+
+  // @ts-ignore
+  @LocalStorage("sidebarHiddenPanels", true, {
+    attribute: false,
+  })
+  private _hiddenPanels: string[] = [];
+
   private _sortable?;
 
   protected render() {
     if (!this.hass) {
       return html``;
     }
-    const debug = false;
-
-    return debug
-      ? html`
-          <ha-sidebar-panel-list
-            .hass=${this.hass}
-            .expanded=${this.expanded}
-            .alwaysExpand=${this.alwaysExpand}
-          ></ha-sidebar-panel-list>
-        `
-      : this.render2();
-  }
-
-  protected render2() {
-    if (!this.hass) {
-      return html``;
-    }
 
     // prettier-ignore
     return html`
-      ${this._renderHeader()} ${this._renderAllPanels()}
+      ${this._renderHeader()}
+      ${this._renderAllPanels()}
       <mwc-list
         attr-for-selected="data-panel"
         @focusin=${this._listboxFocusIn}
@@ -141,7 +274,9 @@ class HaSidebar extends LitElement {
     super.firstUpdated(changedProps);
 
     if (this.hass && this.hass.auth.external) {
-      getExternalConfig(this.hass.auth.external).then(() => {});
+      getExternalConfig(this.hass.auth.external).then((conf) => {
+        this._externalConfig = conf;
+      });
     }
     subscribeNotifications(this.hass.connection, (notifications) => {
       this._notifications = notifications;
@@ -213,19 +348,120 @@ class HaSidebar extends LitElement {
   }
 
   private _renderAllPanels() {
+    // prettier-ignore
     return html`
-      <ha-sidebar-panel-list
-        .hass=${this.hass}
-        .expanded=${this.expanded}
-        .alwaysExpand=${this.alwaysExpand}
-      ></ha-sidebar-panel-list>
+      ${this._renderNormalPanels()}
+      ${this._renderUtilityPanels()}
+      `;
+  }
+
+  private _renderNormalPanels() {
+    const [beforeSpacer] = computePanels(
+      this.hass.panels,
+      this.hass.defaultPanel,
+      this._panelOrder,
+      this._hiddenPanels
+    );
+
+    // prettier-ignore
+    return html`
+      <mwc-list
+        attr-for-selected="data-panel"
+        class="ha-scrollbar"
+        @focusin=${this._listboxFocusIn}
+        @focusout=${this._listboxFocusOut}
+        @scroll=${this._listboxScroll}
+        @keydown=${this._listboxKeydown}
+      >
+        ${this.editMode
+          ? this._renderPanelsEdit(beforeSpacer)
+          : this._renderPanels(beforeSpacer)}
+      </mwc-list>
     `;
   }
 
-  private _renderSpacer(enabled = true) {
-    return enabled
-      ? html`<li divider role="separator" class="spacer"></li>`
-      : html`<li divider role="separator" class="bottom-spacer" disabled></li>`;
+  private _renderPanelsEdit(beforeSpacer: PanelInfo[]) {
+    // prettier-ignore
+    return html`<div id="sortable">
+        ${guard([this._hiddenPanels, this._renderEmptySortable], () =>
+          this._renderEmptySortable ? "" : this._renderPanels(beforeSpacer)
+        )}
+      </div>
+      ${this._renderSpacer()}
+      ${this._renderHiddenPanels()} `;
+  }
+
+  private _renderHiddenPanels() {
+    return html` ${this._hiddenPanels.length
+      ? html`${this._hiddenPanels.map((url) => {
+          const panel = this.hass.panels[url];
+          if (!panel) {
+            return "";
+          }
+          return html`<ha-clickable-list-item
+            @click=${this._unhidePanel}
+            class="hidden-panel"
+            .panel=${url}
+            graphic="icon"
+          >
+            <ha-icon
+              slot="graphic"
+              .icon=${panel.url_path === this.hass.defaultPanel
+                ? "mdi:view-dashboard"
+                : panel.icon}
+            ></ha-icon>
+            <span class="item-text"
+              >${panel.url_path === this.hass.defaultPanel
+                ? this.hass.localize("panel.states")
+                : this.hass.localize(`panel.${panel.title}`) ||
+                  panel.title}</span
+            >
+            <mwc-icon-button class="show-panel">
+              <ha-svg-icon .path=${mdiPlus}></ha-svg-icon>
+            </mwc-icon-button>
+          </ha-clickable-list-item>`;
+        })}
+        ${this._renderSpacer()}`
+      : ""}`;
+  }
+
+  private _renderUtilityPanels() {
+    const [, afterSpacer] = computePanels(
+      this.hass.panels,
+      this.hass.defaultPanel,
+      this._panelOrder,
+      this._hiddenPanels
+    );
+
+    // prettier-ignore
+    return html`
+      <mwc-list
+        @focusin=${this._listboxFocusIn}
+        @focusout=${this._listboxFocusOut}
+        @scroll=${this._listboxScroll}
+        @keydown=${this._listboxKeydown}
+        class="utility-panels"
+      >
+        ${afterSpacer.map((panel) =>
+          this._renderPanel(
+            panel.url_path,
+            panel.url_path === this.hass.defaultPanel
+              ? panel.title || this.hass.localize("panel.states")
+              : this.hass.localize(`panel.${panel.title}`) || panel.title,
+            panel.icon,
+            panel.url_path === this.hass.defaultPanel && !panel.icon
+              ? mdiViewDashboard
+              : undefined
+          )
+        )}
+        ${this._renderExternalConfiguration()}
+        <li divider role="separator" class="spacer"></li>
+      </mwc-list>
+    `;
+  }
+
+  private _renderSpacer() {
+    return html`<li divider role="separator" class="spacer"></li>`;
   }
 
   private _renderNotifications() {
@@ -273,7 +509,6 @@ class HaSidebar extends LitElement {
     return html`<ha-clickable-list-item
       class=${classMap({
         profile: true,
-        // Mimick behavior that mwc-list provides
         "iron-selected": this.hass.panelUrl === "profile",
       })}
       .href=${"profile"}
@@ -295,6 +530,30 @@ class HaSidebar extends LitElement {
         ${this.hass.user ? this.hass.user.name : ""}
       </span>
     </ha-clickable-list-item> `;
+  }
+
+  private _renderExternalConfiguration() {
+    return html`${this._externalConfig && this._externalConfig.hasSettingsScreen
+      ? html`
+          <ha-clickable-list-item
+            aria-role="option"
+            aria-label=${this.hass.localize(
+              "ui.sidebar.external_app_configuration"
+            )}
+            href="#external-app-configuration"
+            tabindex="-1"
+            @click=${this._handleExternalAppConfiguration}
+            @mouseenter=${this._itemMouseEnter}
+            @mouseleave=${this._itemMouseLeave}
+            graphic="icon"
+          >
+            <ha-svg-icon slot="graphic" .path=${mdiCellphoneCog}></ha-svg-icon>
+            <span class="item-text">
+              ${this.hass.localize("ui.sidebar.external_app_configuration")}
+            </span>
+          </ha-clickable-list-item>
+        `
+      : ""}`;
   }
 
   private get _tooltip() {
@@ -334,10 +593,11 @@ class HaSidebar extends LitElement {
     this._sortable = new Sortable(this.shadowRoot!.getElementById("sortable"), {
       animation: 150,
       fallbackClass: "sortable-fallback",
-      // fallbackTolerance: 15,
       dataIdAttr: "data-panel",
       handle: "ha-clickable-list-item",
-      onSort: async () => {},
+      onSort: async () => {
+        this._panelOrder = this._sortable.toArray();
+      },
     });
   }
 
@@ -348,6 +608,38 @@ class HaSidebar extends LitElement {
 
   private _closeEditMode() {
     fireEvent(this, "hass-edit-sidebar", { editMode: false });
+  }
+
+  private async _hidePanel(ev: Event) {
+    ev.preventDefault();
+    const panel = (ev.currentTarget as any).panel;
+    if (this._hiddenPanels.includes(panel)) {
+      return;
+    }
+    // Make a copy for Memoize
+    this._hiddenPanels = [...this._hiddenPanels, panel];
+    this._renderEmptySortable = true;
+    await this.updateComplete;
+    const container = this.shadowRoot!.getElementById("sortable")!;
+    while (container.lastElementChild) {
+      container.removeChild(container.lastElementChild);
+    }
+    this._renderEmptySortable = false;
+  }
+
+  private async _unhidePanel(ev: Event) {
+    ev.preventDefault();
+    const panel = (ev.currentTarget as any).panel;
+    this._hiddenPanels = this._hiddenPanels.filter(
+      (hidden) => hidden !== panel
+    );
+    this._renderEmptySortable = true;
+    await this.updateComplete;
+    const container = this.shadowRoot!.getElementById("sortable")!;
+    while (container.lastElementChild) {
+      container.removeChild(container.lastElementChild);
+    }
+    this._renderEmptySortable = false;
   }
 
   private _itemMouseEnter(ev: MouseEvent) {
@@ -400,8 +692,62 @@ class HaSidebar extends LitElement {
     this._hideTooltip();
   }
 
-  private _listboxKeydown() {
+  private _getIndexOfTarget(evt: Event): number {
+    const listbox = evt.currentTarget as List;
+    const elements = listbox.items;
+    const path = evt.composedPath();
+
+    for (const pathItem of path as Node[]) {
+      let index = -1;
+      if (isNodeElement(pathItem) && isListItem(pathItem)) {
+        index = elements.indexOf(pathItem);
+      }
+
+      if (index !== -1) {
+        return index;
+      }
+    }
+
+    return -1;
+  }
+
+  private _listboxKeydown(ev: KeyboardEvent) {
+    const [beforeSpacer] = computePanels(
+      this.hass.panels,
+      this.hass.defaultPanel,
+      this._panelOrder,
+      this._hiddenPanels
+    );
+
+    const curIndex = this._getIndexOfTarget(ev);
+    const curList = ev.currentTarget as List;
+
+    if (ev.code === "ArrowDown" && curList === this._standardPanelList) {
+      const isLastItem = curIndex === beforeSpacer.length - 1;
+
+      if (isLastItem) {
+        this._setFocusFirstListItem();
+      }
+    } else if (ev.code === "ArrowUp" && curList === this._utilityPanelList) {
+      const isFirstItem = curIndex === 0;
+
+      if (isFirstItem) {
+        this._setFocusLastListItem();
+      }
+    } else if (ev.code === "Enter") {
+      (ev.target as ListItem)?.shadowRoot?.querySelector("a")?.click();
+    }
+
     this._recentKeydownActiveUntil = new Date().getTime() + 100;
+  }
+
+  private _setFocusFirstListItem() {
+    this._utilityPanelList?.focusItemAtIndex(0);
+  }
+
+  private _setFocusLastListItem() {
+    const list = this._standardPanelList;
+    list?.focusItemAtIndex(list?.items.length - 1);
   }
 
   private _showTooltip(item) {
@@ -435,6 +781,13 @@ class HaSidebar extends LitElement {
     fireEvent(this, "hass-show-notifications");
   }
 
+  private _handleExternalAppConfiguration(ev: Event) {
+    ev.preventDefault();
+    this.hass.auth.external!.fireMessage({
+      type: "config_screen/show",
+    });
+  }
+
   private _toggleSidebar(ev: CustomEvent) {
     if (ev.detail.action !== "tap") {
       return;
@@ -442,12 +795,60 @@ class HaSidebar extends LitElement {
     fireEvent(this, "hass-toggle-menu");
   }
 
+  private _renderPanels(panels: PanelInfo[]) {
+    return panels.map((panel) =>
+      this._renderPanel(
+        panel.url_path,
+        panel.url_path === this.hass.defaultPanel
+          ? panel.title || this.hass.localize("panel.states")
+          : this.hass.localize(`panel.${panel.title}`) || panel.title,
+        panel.icon,
+        panel.url_path === this.hass.defaultPanel && !panel.icon
+          ? mdiViewDashboard
+          : undefined
+      )
+    );
+  }
+
+  private _renderPanel(
+    urlPath: string,
+    title: string | null,
+    icon?: string | null,
+    iconPath?: string | null
+  ) {
+    return html`
+      <ha-clickable-list-item
+        .activated=${urlPath === this.hass.panelUrl}
+        .href=${urlPath}
+        data-panel=${urlPath}
+        tabindex="-1"
+        @mouseenter=${this._itemMouseEnter}
+        @mouseleave=${this._itemMouseLeave}
+        graphic="icon"
+      >
+        ${iconPath
+          ? html`<ha-svg-icon slot="graphic" .path=${iconPath}></ha-svg-icon>`
+          : html`<ha-icon slot="graphic" .icon=${icon}></ha-icon>`}
+
+        <span class="item-text">${title}</span>
+        ${this.editMode
+          ? html`<mwc-icon-button
+              class="hide-panel"
+              .panel=${urlPath}
+              @click=${this._hidePanel}
+            >
+              <ha-svg-icon .path=${mdiClose}></ha-svg-icon>
+            </mwc-icon-button>`
+          : ""}
+      </ha-clickable-list-item>
+    `;
+  }
+
   static get styles(): CSSResult[] {
     return [
       haStyleScrollbar,
       css`
         :host {
-          /* height: calc(100% - var(--header-height)); */
           height: 100%;
           display: block;
           overflow: hidden;
@@ -523,26 +924,30 @@ class HaSidebar extends LitElement {
           display: none;
         }
 
-        /* mwc-list.ha-scrollbar {
-          height: 100%;
+        mwc-list.ha-scrollbar {
           --mdc-list-vertical-padding: 4px 0;
           padding: 4px 0;
           display: flex;
+          justify-content: space-between;
           flex-direction: column;
           box-sizing: border-box;
-          height: calc(100% - var(--header-height) - 350px);
+          height: calc(100% - var(--header-height) - 232px);
           height: calc(
-            100% - var(--header-height) - 350px - env(safe-area-inset-bottom)
+            100% - var(--header-height) - 232px - env(safe-area-inset-bottom)
           );
           overflow-x: hidden;
           background: none;
           margin-left: env(safe-area-inset-left);
-          background-color: blue;
-        } */
-
+          -ms-user-select: none;
+          -webkit-user-select: none;
+          -moz-user-select: none;
+          border-right: 1px solid var(--divider-color);
+          background-color: var(--sidebar-background-color);
+          width: 64px;
+        }
         :host([rtl]) mwc-list {
-          margin-left: initial;
-          margin-right: env(safe-area-inset-right);
+          border-right: 0;
+          border-left: 1px solid var(--divider-color);
         }
 
         a {
@@ -552,7 +957,6 @@ class HaSidebar extends LitElement {
           font-size: 14px;
           position: relative;
           outline: 0;
-          border: 1px solid black;
         }
 
         mwc-list-item {
@@ -566,6 +970,7 @@ class HaSidebar extends LitElement {
         :host([expanded]) paper-icon-item {
           width: 248px;
         }
+
         :host([rtl]) mwc-list-item {
           padding-left: auto;
           padding-right: 12px;
@@ -578,6 +983,11 @@ class HaSidebar extends LitElement {
 
         [slot="graphic"] {
           width: 100%;
+        }
+
+        :host([rtl]) mwc-list {
+          margin-left: initial;
+          margin-right: env(safe-area-inset-right);
         }
 
         .iron-selected mwc-list-item::before,
@@ -618,6 +1028,8 @@ class HaSidebar extends LitElement {
           width: 100%;
         }
 
+        a.iron-selected ha-icon,
+        a.iron-selected ha-svg-icon,
         a.iron-selected mwc-list-item ha-icon,
         a.iron-selected mwc-list-item ha-svg-icon {
           color: var(--sidebar-selected-icon-color);
@@ -636,14 +1048,6 @@ class HaSidebar extends LitElement {
           display: block;
         }
 
-        .divider {
-          padding: 10px 0;
-        }
-        .divider::before {
-          display: block;
-          height: 100px;
-          background-color: var(--divider-color);
-        }
         .notifications-container {
           display: flex;
           margin-left: env(safe-area-inset-left);
@@ -680,7 +1084,7 @@ class HaSidebar extends LitElement {
         }
 
         .notification-badge {
-          /* min-width: 20px;
+          min-width: 20px;
           box-sizing: border-box;
           border-radius: 50%;
           font-weight: 400;
@@ -689,7 +1093,7 @@ class HaSidebar extends LitElement {
           text-align: center;
           padding: 0px 4px;
           color: var(--text-accent-color, var(--text-primary-color));
-          font-size: 14px; */
+          font-size: 14px;
         }
         ha-svg-icon + .notification-badge {
           position: absolute;
@@ -701,12 +1105,6 @@ class HaSidebar extends LitElement {
         .spacer {
           flex: 1;
           pointer-events: none;
-        }
-
-        .bottom-spacer {
-          flex: 1;
-          pointer-events: none;
-          height: 100%;
         }
 
         .subheader {
