+++ src/layouts/home-assistant-main.ts
@@ -1,4 +1,3 @@
-import "../components/ha-sidebar-combined";
 import "@polymer/app-layout/app-drawer-layout/app-drawer-layout";
 import type { AppDrawerLayoutElement } from "@polymer/app-layout/app-drawer-layout/app-drawer-layout";
 import "@polymer/app-layout/app-drawer/app-drawer";
@@ -86,13 +85,13 @@ class HomeAssistantMain extends LitElement {
           .persistent=${!this.narrow &&
           this.hass.dockedSidebar !== "always_hidden"}
         >
-          <ha-sidebar-combined
+          <ha-sidebar
             .hass=${hass}
             .narrow=${sidebarNarrow}
             .editMode=${this._sidebarEditMode}
             .alwaysExpand=${sidebarNarrow ||
             this.hass.dockedSidebar === "docked"}
-          ></ha-sidebar-combined>
+          ></ha-sidebar>
         </app-drawer>
 
         <partial-panel-resolver
