--- src/resources/ha-sortable-style-mwc.ts
@@ -1,99 +1,0 @@
-import { css } from "lit-element";
-
-export const sortableStyles = css`
-  #sortable mwc-list-item:nth-of-type(2n) {
-    animation-name: keyframes1;
-    animation-iteration-count: infinite;
-    transform-origin: 50% 10%;
-    animation-delay: -0.75s;
-    animation-duration: 0.25s;
-  }
-
-  #sortable mwc-list-item:nth-of-type(2n-1) {
-    animation-name: keyframes2;
-    animation-iteration-count: infinite;
-    animation-direction: alternate;
-    transform-origin: 30% 5%;
-    animation-delay: -0.5s;
-    animation-duration: 0.33s;
-  }
-
-  #sortable mwc-list-item {
-    height: 48px;
-    display: flex;
-  }
-
-  #sortable {
-    outline: none;
-    display: block !important;
-  }
-
-  .hidden-panel {
-    display: flex !important;
-  }
-
-  .sortable-fallback {
-    display: none;
-  }
-
-  .sortable-ghost {
-    opacity: 0.4;
-  }
-
-  .sortable-fallback {
-    opacity: 0;
-  }
-
-  @keyframes keyframes1 {
-    0% {
-      transform: rotate(-1deg);
-      animation-timing-function: ease-in;
-    }
-
-    50% {
-      transform: rotate(1.5deg);
-      animation-timing-function: ease-out;
-    }
-  }
-
-  @keyframes keyframes2 {
-    0% {
-      transform: rotate(1deg);
-      animation-timing-function: ease-in;
-    }
-
-    50% {
-      transform: rotate(-1.5deg);
-      animation-timing-function: ease-out;
-    }
-  }
-
-  .show-panel,
-  .hide-panel {
-    display: none;
-    position: absolute;
-    top: 0;
-    right: 0;
-    --mdc-icon-button-size: 40px;
-  }
-
-  .hide-panel {
-    top: 4px;
-    right: 8px;
-  }
-
-  :host([expanded]) .hide-panel {
-    display: block;
-  }
-
-  :host([expanded]) .show-panel {
-    display: inline-flex;
-  }
-
-  mwc-list-item.hidden-panel,
-  mwc-list-item.hidden-panel span,
-  mwc-list-item.hidden-panel ha-icon[slot="item-icon"] {
-    color: var(--secondary-text-color);
-    cursor: pointer;
-  }
-`;
